feat: Add @react-navigation/bottom-tabs dependency
This commit adds the "@react-navigation/bottom-tabs" package as a dependency in both "package.json" and "package-lock.json". The version added is "^6.5.12". This package is required for creating bottom tab navigation in the app.

The commit also introduces a new file, "BottomTabsStack.tsx", which contains the implementation of the bottom tab navigator using the newly added dependency. It defines two screens, "Home" and "Profile", and sets up initial routing configurations.

Additionally, there are some minor changes to other files such as updating import statements and modifying existing code to use the new bottom tab navigator.

Closes #123